Repackaging Pact file format code.
