+++ scalapact-core/src/main/scala/com/itv/scalapactcore/common/matchir/IrNodeRule.scala
@@ -1,7 +1,7 @@
 package com.itv.scalapactcore.common.matchir
 
-import com.itv.scalapactcore.MatchingRule
 import com.itv.scalapactcore.common.matchir.PactPathParseResult.{PactPathParseFailure, PactPathParseSuccess}
+import com.itv.scalapactcore.common.pact.MatchingRule
 
 import scala.util.Random
