+++ scalapact-core/src/test/scala/com/itv/scalapactcore/ScalaPactReaderWriterSpec.scala
@@ -1,28 +1,28 @@
 package com.itv.scalapactcore
 
 import org.scalatest.{FunSpec, Matchers}
-
 import argonaut.JsonParser._
+import com.itv.scalapactcore.common.pact.{PactReader, PactWriter}
 
 class ScalaPactReaderWriterSpec extends FunSpec with Matchers {
 
   describe("Reading and writing a homogeneous Pact files") {
 
     it("should be able to read Pact files") {
-      val pactEither = ScalaPactReader.jsonStringToPact(PactFileExamples.simpleAsString)
+      val pactEither = PactReader.jsonStringToPact(PactFileExamples.simpleAsString)
 
       pactEither.right.get shouldEqual PactFileExamples.simple
     }
 
     it("should be able to read Pact files using the old provider state key") {
-      val pactEither = ScalaPactReader.jsonStringToPact(PactFileExamples.simpleOldProviderStateAsString)
+      val pactEither = PactReader.jsonStringToPact(PactFileExamples.simpleOldProviderStateAsString)
 
       pactEither.right.get shouldEqual PactFileExamples.simple
     }
 
     it("should be able to write Pact files") {
 
-      val written = ScalaPactWriter.pactToJsonString(PactFileExamples.simple)
+      val written = PactWriter.pactToJsonString(PactFileExamples.simple)
 
       val expected = PactFileExamples.simpleAsString
 
@@ -31,11 +31,11 @@ class ScalaPactReaderWriterSpec extends FunSpec with Matchers {
 
     it("should be able to eat it's own dog food") {
 
-      val json = ScalaPactWriter.pactToJsonString(PactFileExamples.simple)
+      val json = PactWriter.pactToJsonString(PactFileExamples.simple)
 
-      val pact = ScalaPactReader.jsonStringToPact(json).right.get
+      val pact = PactReader.jsonStringToPact(json).right.get
 
-      val `reJson'd` = parse(ScalaPactWriter.pactToJsonString(pact)).toOption.get
+      val `reJson'd` = parse(PactWriter.pactToJsonString(pact)).toOption.get
 
       `reJson'd` shouldEqual parse(PactFileExamples.simpleAsString).toOption.get
 
@@ -44,14 +44,14 @@ class ScalaPactReaderWriterSpec extends FunSpec with Matchers {
     }
 
     it("should be able to read ruby format json") {
-      val pactEither = ScalaPactReader.jsonStringToPact(PactFileExamples.simpleAsString)
+      val pactEither = PactReader.jsonStringToPact(PactFileExamples.simpleAsString)
 
       pactEither.right.get shouldEqual PactFileExamples.simple
     }
 
     it("should be able to write a pact file in ruby format") {
 
-      val written = ScalaPactWriter.pactToJsonString(PactFileExamples.simple)
+      val written = PactWriter.pactToJsonString(PactFileExamples.simple)
 
       val expected = PactFileExamples.simpleAsString
 
@@ -61,11 +61,11 @@ class ScalaPactReaderWriterSpec extends FunSpec with Matchers {
 
     it("should be able to eat it's own dog food with no body") {
 
-      val json = ScalaPactWriter.pactToJsonString(PactFileExamples.verySimple)
+      val json = PactWriter.pactToJsonString(PactFileExamples.verySimple)
 
-      val pact = ScalaPactReader.jsonStringToPact(json).right.get
+      val pact = PactReader.jsonStringToPact(json).right.get
 
-      val `reJson'd` = parse(ScalaPactWriter.pactToJsonString(pact)).toOption.get
+      val `reJson'd` = parse(PactWriter.pactToJsonString(pact)).toOption.get
 
       `reJson'd` shouldEqual parse(PactFileExamples.verySimpleAsString).toOption.get
 
@@ -74,14 +74,14 @@ class ScalaPactReaderWriterSpec extends FunSpec with Matchers {
     }
 
     it("should be able to read ruby format json with no body") {
-      val pactEither = ScalaPactReader.jsonStringToPact(PactFileExamples.verySimpleAsString)
+      val pactEither = PactReader.jsonStringToPact(PactFileExamples.verySimpleAsString)
 
       pactEither.right.get shouldEqual PactFileExamples.verySimple
     }
 
     it("should be able to write a pact file in ruby format with no body") {
 
-      val written = ScalaPactWriter.pactToJsonString(PactFileExamples.verySimple)
+      val written = PactWriter.pactToJsonString(PactFileExamples.verySimple)
 
       val expected = PactFileExamples.verySimpleAsString
