+++ scalapact-core/src/test/scala/com/itv/scalapactcore/common/matchir/IrNodeRuleSpec.scala
@@ -1,6 +1,6 @@
 package com.itv.scalapactcore.common.matchir
 
-import com.itv.scalapactcore.MatchingRule
+import com.itv.scalapactcore.common.pact.MatchingRule
 import org.scalatest.{FunSpec, Matchers}
 
 import scala.language.postfixOps
