+++ scalapact-core/src/test/scala/com/itv/scalapactcore/pactspec/util/PactSpecTester.scala
@@ -1,8 +1,8 @@
 package com.itv.scalapactcore.pactspec.util
 
-import com.itv.scalapactcore.{Interaction, InteractionRequest, InteractionResponse}
 import com.itv.scalapactcore.common.matching.InteractionMatchers._
 import com.itv.scalapactcore.common.matching.{MatchOutcomeFailed, MatchOutcomeSuccess}
+import com.itv.scalapactcore.common.pact.{Interaction, InteractionRequest, InteractionResponse}
 import org.scalatest.{FunSpec, Matchers}
 
 trait PactSpecTester extends FunSpec with Matchers {
@@ -24,10 +24,6 @@ trait PactSpecTester extends FunSpec with Matchers {
       val i = Interaction(None, None, "", spec.expected, InteractionResponse(None, None, None, None))
 
       mode match {
-//        case NonStrictOnly =>
-//          doRequestMatch(spec, i, strictMatching = false, shouldMatch = spec.`match`, path)
-//          doRequestMatch(spec, i, strictMatching = true, shouldMatch = !spec.`match`, path)
-
         case StrictOnly =>
           doRequestMatch(spec, i, strictMatching = false, shouldMatch = !spec.`match`, path)
           doRequestMatch(spec, i, strictMatching = true, shouldMatch = spec.`match`, path)
@@ -107,5 +103,6 @@ sealed trait StrictTestMode
 // Strict means 'Pact standard' and non-strict is a relaxing of the rules.
 // There should be no permissive only tests
 //case object NonStrictOnly extends StrictTestMode
+
 case object StrictOnly extends StrictTestMode
 case object StrictAndNonStrict extends StrictTestMode
